+++ examples/animation.rs
@@ -65,11 +65,10 @@ impl State for GameState {
     }
 }
 
-fn main() -> tetra::Result {
+fn main() {
     ContextBuilder::new("Displaying an Animation", 64, 64)
         .maximized(true)
         .resizable(true)
         .quit_on_escape(true)
-        .build()?
-        .run_with(GameState::new)
+        .run_with(GameState::new);
 }