+++ examples/audio.rs
@@ -80,9 +80,8 @@ impl State for GameState {
     }
 }
 
-fn main() -> tetra::Result {
+fn main() {
     ContextBuilder::new("Audio Playback", 640, 480)
         .quit_on_escape(true)
-        .build()?
-        .run_with(GameState::new)
+        .run_with(GameState::new);
 }