+++ examples/canvas.rs
@@ -78,11 +78,10 @@ impl State for GameState {
     }
 }
 
-fn main() -> tetra::Result {
+fn main() {
     ContextBuilder::new("Canvas Rendering", 1280, 720)
         .maximized(true)
         .resizable(true)
         .quit_on_escape(true)
-        .build()?
-        .run_with(GameState::new)
+        .run_with(GameState::new);
 }