+++ examples/hello_world.rs
@@ -11,8 +11,6 @@ impl State for GameState {
     }
 }
 
-fn main() -> tetra::Result {
-    ContextBuilder::new("Hello, world!", 1280, 720)
-        .build()?
-        .run(&mut GameState)
+fn main() {
+    ContextBuilder::new("Hello, world!", 1280, 720).run(GameState);
 }