+++ examples/interpolation.rs
@@ -97,12 +97,11 @@ impl State for GameState {
     }
 }
 
-fn main() -> tetra::Result {
+fn main() {
     ContextBuilder::new("Interpolation", 640, 480)
         .tick_rate(5.0)
         .maximized(true)
         .resizable(true)
         .quit_on_escape(true)
-        .build()?
-        .run_with(GameState::new)
+        .run_with(GameState::new);
 }