+++ examples/scaling.rs
@@ -67,11 +67,10 @@ impl State for GameState {
     }
 }
 
-fn main() -> tetra::Result {
+fn main() {
     ContextBuilder::new("Screen Scaling", 640, 480)
         .resizable(true)
         .maximized(true)
         .quit_on_escape(true)
-        .build()?
-        .run_with(GameState::new)
+        .run_with(GameState::new);
 }