+++ src/platform.rs
@@ -21,7 +21,7 @@ use crate::input::{self, GamepadAxis, GamepadButton, Key};
 use crate::window;
 use crate::{Context, ContextBuilder};
 
-pub struct SdlPlatform {
+pub struct Platform {
     sdl: Sdl,
     window: Window,
 
@@ -46,8 +46,8 @@ struct SdlController {
     slot: usize,
 }
 
-impl SdlPlatform {
-    pub fn new(builder: &ContextBuilder<'_>) -> Result<(SdlPlatform, GlContext, i32, i32)> {
+impl Platform {
+    pub fn new(builder: &ContextBuilder<'_>) -> Result<(Platform, GlContext, i32, i32)> {
         let sdl = sdl2::init().map_err(TetraError::Sdl)?;
 
         let video_sys = sdl.video().map_err(TetraError::Sdl)?;
@@ -131,7 +131,7 @@ impl SdlPlatform {
             .gl_set_swap_interval(if builder.vsync { 1 } else { 0 })
             .map_err(TetraError::Sdl)?;
 
-        let platform = SdlPlatform {
+        let platform = Platform {
             sdl,
             window,